Remove specialised horizontal line hardware blit
Hardware blitting of a 1-pixel width horizontal line was done by
a separate function for marginal performance benefits.  However, the
code did not support multifill, which is valid for horizontal lines
drawn via linea4.  We now use the same code in this situation as for
other widths of horizontal line or horizontal rectangles.