Fix problem with Hatari in extended VDI modes
If EmuTOS was started using extended VDI modes with 4 planes, the
screen display was messed up.  This was because Hatari had hacked
v_planes but, as of commit 690ddb2, EmuTOS uses the variable
v_planes_shift for screen address calculations, and that variable
had not been changed.  Thanks to Eero Tamminen for pointing this out.