Clear BSS/heap for programs without relocations
Atari TOS 1.04 and above clears the BSS or the entire heap (depending
on the header flags) for programs that have the 'absflag' field in
the program header set, indicating that there is no relocation table.

This commit fixes the text editor Clix which previously crashed or
failed with an error message due to its BSS not being zeroed.