Force media change for floppy disks when pressing "Esc"
Atari TOS forces a media change and subsequent re-read of a disk
when the user presses "Esc" on the Desktop. EmuTOS now does the same
for floppy disks, using an undocumented feature of Rwabs().
For non-floppies it's still the responsibility of the hardware/driver
to return the appropriate media change status.