Fix palette value returned by v_opnvwk()
The value returned by v_opnvwk() in work_out[39] was always 2.  It
should (nominally) be the number of different colours available in
the current resolution.  This caused problems with the COLOR.CPX
distributed on the MegaSTe language disk, which used this value to
determine whether to enable its colour sliders.  Thanks to Matthias
Arndt for reporting the problem, and to Christian Zietz and Thorsten
Otto for their debugging.