Remove sectors/track check in floppy formatter
Unlike Atari TOS, EmuTOS returned an error from Flopfmt() if an
attempt was made to write more than 10 sectors/track to a double
density drive.  However, there are HD floppy add-ons for the ST
which do automatic density switching and work correctly under Atari
TOS, but do not work under EmuTOS since it is not aware that an HD
drive is installed, and so Flopfmt() returns an error code.  This
commit removes the check for sectors/track based on drive type.