Simplify lowlevel directory search
The function findit() returned either a negative error number or NULL
to signal an error.  Since the callers of findit() always check for
NULL & return an appropriate error code depending on the current
function, findit() now only returns NULL for an error.  This saves
about 48 bytes in the 192K ROMs.