Fix EmuCON copy/move command
Copying a file to a file (rather than a folder) always failed with
the message "<target file> is not a directory".  This happened
whether or not <target file> existed.  This error was introduced
by commit 5695e1c7e183c264010551c9b3110703f386f800.
Thanks to Vincent Riviere for reporting this.