Fix delay_loop()
Since the '_count' variable was only specified as an input to
delay_loop(), the compiler was allowed to assume that it remained
unchanged. If delay_loop() was used multiple times near each other,
the count-down register was in some case not reinitialized,
causing all but the first delay_loop to be without effect.

The impact was mitigated by the (unnecessary) 'memory' clobber
in delay_loop() which forced the compiler to reload the count-down
variable if it was not a local variable but stored in memory.

Thank you to Roger Burrows for reporting this.