Fix key repeat handling
If key repeat was in progress, and a second key was pressed and held
down then the second key would start repeating instead of the first
one.  This is the same as Atari TOS.  However, if the first key was
then released, key repeat would stop, which is wrong: only the
release of the currently-repeating key should stop the repeat.  This
problem also affected mouse movement via the alt-arrow keys.
Thanks to Chris (exxos) for reporting the problem.