Fix bug in keyboard mouse emulation
If keyboard mouse emulation was active, with Shift/Alt/arrowkey down,
and the arrowkey was released, mouse emulation mode was not exited
properly.  As a result, if you then toggled the Shift key, keeping
the Alt key down, the mouse cursor jumped 8 pixels at a time.  This
bug was introduced by commit 988c620.