Support buggy FPU decoding on PAK68/2 accelerator
The PAK68/2, a 68020 accelerator for 68000-based systems, does
not take into account the coprocessor ID field in Line F
instructions when generating the chip select signal for an
optional FPU. Therefore, PMMU instructions are erroneously sent
to the FPU, too.

Because of this, the previous implementation caused an unhandled
illegal instruction exception on the PAK68/2. The new choice of
PMMU instructions prevents this.

Thank you to Frank Lukas for reporting this.