Fix obscure bug in EmuDesk 'Open folder'
Under EmuTOS, if the Alt key was held down when opening a folder on
the desktop, the folder was opened in a new window.  Under Atari
TOS, this was only true if no other modifier key was held down at
the same time.  EmuTOS now behaves the same as Atari TOS.