Fix sound mixer initialization for TT/STE
Previously, when programming the LMC1992 mixer via the Microwire
interface, EmuTOS assumed that the first bit would be shifted
within ca. 2 µs. However, measurements on a TT show that it can
take over 10 µs before a Microwire transfer is even started. Thus,
the previous implementation would detect a transfer as finished
when it had not been started.

Now, the same mechanism is used as in Atari TOS: utilizing the LSB
of the data word as 'stop' bit and waiting until the data register
becomes zero.