Fix handling of window arrow gadgets
If a button was already down, the ctlmgr() function called the
dispatcher to allow tasks to run, rather than calling evnt_multi().
This meant that button events could be detected by tasks before the
ctlmgr() intercepted them and converted them to system messages.
This happened, for example, if a task's evnt_multi() was waiting for
a button down event as well as messages.