Improve display of EmuDesk function-key assignments
Previously, single-digit function key numbers were always displayed
as two characters (with or without a leading zero).  For example, F1
was displayed as either "F 1" or "F01".  Now single-digit numbers
are displayed as one character, e.g. F9 is displayed as "F9".  The
display of two-digit function key numbers is unchanged.  Thanks to
Vincent Riviere for suggesting the change.