Add optional cold-boot delay
Optionally, EmuTOS can now wait a user-configurable time (up to 255 seconds)
at cold-boot before trying to access any disks. This is useful particularly
in the case of hard disks / hard disk replacements that are powered on
at the same time as the computer, e.g. an internal Ultrasatan. The delay
allows these disks to become ready before EmuTOS accesses them.

Since many supported systems do not have non-volatile RAM to store this
setting, the boot delay is stored in the EmuTOS ROM, in a new field inside
the OSEX header: An unsigned byte at offset 0x4C from the beginning of the
ROM contains the delay in seconds. With the default value of 0, there is no
additional wait time. The user is supposed to utilize a hex editor or the
supplied commandline tool (tools/boot-delay.c) before using the ROM.