Fix incompatibility with environment strings
In all versions of Atari TOS, the process that launches the programs
in the AUTO folder has an empty environment string, whereas in EmuTOS
the process did not have a valid environment string (it had a null
pointer in the PD).  This caused a crash in ACCFOLDR.PRG when it
looked for the empty string in order to replace it.