Add white-list checking for Atari partition ids
Previously, any partition id in the standard Atari partition table
was accepted as valid and, if also marked as 'available', caused
a disk letter to be reserved.  However, HDDRIVER validates partition
ids against a white-list, which caused differences in behaviour for
partition ids 'ACK' & 'OOP' which are used by Spectre GCR.
Now we use a white-list too (we already did this for partitions in
the ICD/Supra extended partition table).