Falcon: Switch off STe bus mode as early as possible
After boot, the Atari Falcon is in a mode where most bus
errors are disabled, which affects hardware detection by
EmuTOS. Particularly, on a 4 MB Falcon, a Nova graphics
card adapter was erroneously detected. Now, the bus errors
are enabled as soon as EmuTOS determines that it runs on a
Falcon.

Thank you to Pavel Salač for reporting this issue and to
Roger Burrows for in-depth testing of the effects of
"STe bus emulation" mode.