Reset FPU before any further use
On the Atari TT the FPU is not reset by the CPU's RESET instruction.
Previously, if there was an FPU exception pending after a keyboard
reset, the exception would be triggered by FNOP, which caused EmuTOS
to hang at boot. Now, the first FPU instruction is an FRESTORE that
resets the FPU, clearing and disabling exceptions.