Fix bug setting ST resolutions on the Falcon
On the Falcon, when a Setscreen() is done for an ST resolution, we
now update the current videl mode to correspond to the new ST
resolution.  This ensures that a subsequent VsetRGB() will update
the STe palette registers rather than the Falcon palette registers.
This fixes a problem with DELMPAINT reported by Eero Tamminen.