Align memory allocations to 4 bytes on Falcon
Atari TOS 1 - 3 aligns memory requested from ST-RAM on multiples of
two bytes, whereas Atari TOS 4 aligns on multiples of four bytes.
As programs might (implicitly) rely on this detail, EmuTOS bases
its alignment on the type of machine it is running on. Note:
Alt-RAM memory blocks are always aligned on multiples of 4 bytes.

This fixes "Les Dinosaures". Thank you to Eero Tamminen for reporting
this issue.