Add early EmuTOS startup message
We now display a simple startup message as soon as possible to let
the user know that we're alive.  This is especially important when
booting on a system with no hard disk and no inserted diskette; in
such a case, there is otherwise a long period with a blank screen.