Fix bug with desktop redisplay after program exit
This fixes a problem that happens occasionally when DELMPAINT exits:
the desktop needs an extra mouse button click to display properly.
This could happen with any program that uses the AES, but is most
likely to happen if the program uses the mouse a lot, but only
calls the AES occasionally.