Fix display bug in desktop copy/move
When you select OK for a copy/move operation, the OK button should
remain selected until the end of the operation.  However, if the
'name conflict' dialog was displayed during the copy/move, the OK
button of the copy/move dialog was deselected for the rest of the
copy.  Thanks to Miro Kropacek for pointing this out.