Fix bug in vq_color()
When doing a vq_color(), the VDI pen number is mapped to a hardware
palette register.  This value was not clamped according to the
current number of colours, causing the wrong values to be returned
on a system with Videl.  This caused problems when running Interface,
which uses vq_color()/vs_color() to save & restore the palette.
The bug did not show up until commit 26197e1, which applied the
corresponding (necessary) clamping within vs_color().