Add optional debug output to the cartridge port
This can be enabled by setting CARTRIDGE_DEBUG_PRINT to 1.

A character 'c' is encoded into address lines A8-A1 by performing
a read access to the upper half of the cartridge address space
0xFB0xxx, which is decoded to the /ROM3 signal. It has the advantage
of being always available since it does not require initialization of
a peripheral.

Output can be decoded by connecting a logic analyzer to A8-A1,
triggering on /ROM3, or by using a special firmware for the
SidecarTridge Multi-device hardware:
https://github.com/czietz/atari-debug-cart