Fix cosmetic desktop bug when dragging an icon
When a desktop icon is dragged, each icon that it is dragged over
is highlighted.  This is unlike Atari TOS, which only highlights
icons that represent potential targets: thus it does not highlight
icons for non-executable files.  This discrepancy is now fixed.
Thanks to Vincent Riviere for pointing this out.