Centralise localisation info
In preparation for splitting the 512K ROMs to provide room for
future enhancements, we introduce a single file (localise.ctl)
that combines localisation-related info from country.mk and
po/LINGUAS.  This is processed by a new program (localise) to
generate required files for the build process.  The actual
split of the ROMs will be implemented in a future commit.