Fix build error for UNIQUE builds
If you start from a fresh download and build a country-specific
version of EmuTOS, it will fail because obj/header.h does not exist.
This is ultimately because makefile.dep depends on $(GEN_SRC), which
includes util/langs.c which make does not know how to build.  This
causes make to ignore other dependencies in $(GEN_SRC), including
obj/header.h.  This problem was introduced by commit 51725b0.
Many thanks to Christian Zietz for debugging this.