Nova/Vofa: Make Physbase() return VGA memory address
Like the Nova driver EMULATOR.PRG, EmuTOS now returns the address
of the memory on the VGA card as physical start address of the
video memory. This fixes software that directly accesses the video
memory.

Thank you to Jo Even Skarstein for reporting this issue for the
Pure C editor.