Reset column counter when entering text mode
The VT-52 escape sequence used when entering text mode advances the
BDOS column counter. Therefore, a TOS program would get incorrect tab
expansion in the first output line. An additional carriage return
ensures that the column counter is reset to zero and is otherwise
harmless since the cursor is at the beginning of the first line, anyway.