Restore 'reserved' word in font header structure
This is part of the official Atari definition of the font header
(see p J-8 of "GEM Programmer's Guide, Vol.1: VDI", Third Edition,
January 1989).  It was removed erroneously by commit e477334.
Note: since a copy of this is contained within the virtual
workstation structure, the length of this structure affects the
offsets of certain workstation variables.  Warp9 (at least) cares
about this.