Remove CONF_SERIAL_CONSOLE_INTERRUPT_MODE
Interrupt mode is the default when CONF_SERIAL_CONSOLE is enabled.

Set CONF_SERIAL_CONSOLE_POLLING_MODE to 1 to use polling mode instead.
This is necessary when ikbdiorec is not filled on interrupt
when CONF_SERIAL_CONSOLE is enabled.
This can be useful in early stages when porting EmuTOS to new hardware
and bconstat2()/bconin1() are implemented by polling.

FIXME: This breaks CONF_SERIAL_CONSOLE on non-ColdFire targets.