Fix EmuDesk handling of menu item enable/disable
For compatibility with Atari TOS, the 'File|Open', 'File|Delete' and
'Options|Install application' menu items are now enabled if there
are selected files/folders in a window, even if none are currently
visible due to window scrolling.  Previously at least one had to be
visible to enable the menu items.  Thanks to Vincent Barilliot for
pointing out this difference.