Increase serial port buffer size to 256
Since commit 0eff02d (which enabled interrupt-driven input), the
input buffer for MFP#1 (ST-compatible) is required.  The previous
size caused problems with some applications; the size is now the
same as Atari TOS.  Note that this change increases the size of all
serial port buffers, including those not (yet) in use.