Fix EmuDesk crash when double-clicking desktop
If you double-click on the desktop background while a window is
open, EmuDesk crashes in do_open().  This happens because it passes
an obid of zero to i_find(), which passes it to app_afind_by_id(),
which returns the address of a 'random' anode with no corresponding
screen object.  Thanks to Christian Zietz for finding the offending
commit.  Fixed by checking for a valid obid in app_afind_by_id().