Fix 'slider creep' in AES windows
If you single click on the vertical or horizontal slider in a window,
it will often move upwards/leftwards even if you haven't moved the
mouse.  This is because the translation from slider position to pixel
position and back again does not yield the original result, because
the mul_div() function does not round.  This bug was introduced by
commit f891fd7 (Nov/2002).  We now call mul_div_round() instead.