Improve AES window sizing via SIZER gadget
Sizing a window via the use of the SIZER gadget has been improved in
two respects.  Firstly, the presence of gadgets in the title bar now
affects the minimum width in the same way as the presence of gadgets
in the horizontal bar.  Secondly, the minimum size in each direction
has been reduced.  Sizing is now somewhat more like Atari TOS 4.