Make EmuDesk show busy while setting default path
During folder & application open, EmuDesk sets the default path for
a drive.  If the medium has (or may have) changed, this requires disk
i/o.  This is particularly noticeable on a write-protected floppy
because, due to the method required to detect media change, such
floppies are almost always in a "medium may have changed" state.

We now show the busy mouse cursor during this function.  This fixes
the problem where there is a noticeable delay between double-clicking
on a file/folder and the busy cursor appearing.  Thanks to exxos
for pointing out this problem.