Fix warning with GCC 11.1.0
util/cookie.c: In function 'get_frb_cookie':
util/cookie.c:104:16: warning: 'frbvalue' may be used uninitialized [-Wmaybe-uninitialized]
  104 |         return frbvalue;
      |                ^~~~~~~~
util/cookie.c:101:12: note: 'frbvalue' declared here
  101 |     UBYTE *frbvalue;
      |            ^~~~~~~~

This is because of the strict-aliasing rule.
GCC can't see that a variable can be modified when a pointer to it
is passed to a function using a cast to another pointer type.