Disable 3D objects for 256K ROMs
This is for compatibility with Atari TOS 2, which is the target for
the 256K EmuTOS ROMs.  Specifically, this avoids problems with GFA
Basic: if launched in low-res, it will switch to medium-res without
notifying the AES, and it can then call the file selector or issue
a form alert.  The dialogs will be set up with backgrounds (& other
objects) in a colour that is invalid in medium-res, and will be
displayed as black.  Since this trick worked fine in early versions
of Atari TOS, other programs may also rely on it working.