Increase Falcon screen memory allocation
The current memory allocation allows for a maximum screen size of
640x480x256; however the screen size can be 768x480x256 when overscan
is set.  This mismatch causes the end of screen memory to be mapped
into non-existant RAM.  Under Hatari, this causes a crash during
startup; on real hardware it causes the bottom of the screen to
contain garbage.  Thanks to David Henderson for finding this.