Increase timeout for IDE transfers
Hard drives that enter power-saving mode might need more than the previous
timeout of 1 second to complete a transfer. In particular, this change
fixes issues with the ST1 hard drive (in CF form factor) that uses a very
aggressive power-saving scheme and that (according to its datasheet) can
take up to 2.5 seconds to transition from standby to ready.