Rework detection of partitionless disks
The previous method to check if a hard disk does not have a DOS MBR (incl.
partition table) but only a boot sector relied on the presence of a "magic"
string ("FAT16"). However, it is not mandatory for a FAT boot sector to
contain that string. Thus, the new detection routine now uses a different
heuristic to tell a boot sector apart from an MBR.