Add initial support for extended object types
The AES now explicitly supports MagiC-style underlining of a G_STRING
object: if the ob_state has WHITEBAK set and the upper byte set to
0xFF, the object will be underlined.  This is controlled by the
CONF_WITH_EXTENDED_OBJECTS config variable, which should be used for
control of future extended object support.  Thanks to Thorsten Otto
for providing info on this feature.