Fix bug in vqt_name()
The font index passed to vqt_name() is an index to the font family:
all fonts with the same id belong to one family.  vqt_name() was
treating each font as a separate family.  This bug was introduced
by commit 98f91b0 (may/2002) and mostly affects systems using GDOS.
Thanks to 'tevorbl' on GitHub for pointing this out.