Revert commit 960ce65
This set the AES version to 3.30 if CONF_WITH_WINDOW_COLOURS was
enabled.  However, setting this caused a crash in DevPac 3.10.
As a result of this change, the WCOLORS CPX will no longer work
properly because it is not aware that wind_set(WF_COLOR/WF_DCOLOR)
is available.  Also, programs will not be able to detect the
availability of 3D object support (introduced in AES 3.40).