Fix bug in setting screen address from _screenpt
If _screenpt ($45e) is nonzero, the VBL uses it to set the physical
screen address.  This did not work correctly on non-ST hardware when
setting an address that was not on a 256-byte boundary, because the
code updated the hardware registers in the wrong order.  Thanks to
Vincent Riviere for pointing this out.