Tweak submenu positioning slightly
As per Atari documentation, we already move a submenu to the left of
the associated main menu item if it would otherwise extend past the
right hand edge of the screen.  This commit implements the additional
rule: if the above would cause the submenu to extend past the left
hand edge of the screen, it is moved right a character width at a
time until that is no longer the case.