Fix obscure EmuDesk bug in 'Info/Rename'
This bug is quasi-random but is easiest to reproduce just after
EmuDesk loads.  Open a disk, select a file/folder, do an Info/Rename:
the info of the file/folder is shown.  Press Cancel: the file/folder
(correctly) remains selected.  Do an Info/Rename again: now the info
for the entire disk is shown.
Cause: the function fun_rebld_marked() follows the chain of WNODEs
and rebuilds windows that have the WN_REBUILD flag set.  However, it
did not check that the WNODE was valid (w_id==0).  Since the flags
field is not initialised, residual junk could cause rebuild_window()
to be called with w_id=0, causing desk_verify() to reset some global
variables incorrectly.  This bug was introduced by commit f93e029f.