Make FireBee ROMs single-language
A current multi-language version of EmuTOS requires more than 512K,
mostly due to support of additional languages.  If you try to load
such an image via the FireTOS boot prompt, EmuTOS will crash due to
a bug in FireTOS.  Limiting EmuTOS to a single language reduces the
image size to below 350K and circumvents the crash.