Tweak EmuDesk horizontal window alignment
EmuDesk windows are normally snapped horizontally to the nearest
16-pixel boundary.  If a window was partially moved offscreen to the
right, and placed such that it was further snapped rightwards, the
mover bar could become inaccessible, and the window became stuck.
This could occur in low resolutions with 3D objects enabled.  In this
situation, we now move the window leftwards by 16 pixels.