Fix prg256 target
The syntax below was problematic:
prg256: $(eval UNIQUE := $(COUNTRY))

Usage of $(eval ...) actually changed the UNIQUE value for:
- *all* the rules in the *whole* Makefile,
  just as if the affectation was written before the rule.
- all the ifneq tests potentially *after* that line (there was none)

Unlike expected, this didn't only affect the current rule.
It seems that the only affected file was obj/country with the 1024 target,
and that was harmless.

Note that invoking recursive $(MAKE) is *mandatory*
when some variable used by ifneq (like UNIQUE) is overridden by the rule.

When invoking recursive $(MAKE), it is also *mandatory*
to add the current rule to the NODEP list
so the "draft" tool is run after the recursion,
effectively using the overridden values.
This was lacking for prg256, causing usage of full 512k resources.