ColdFire: Compute SDCLK freq at run time instead of compile time
The code in bios/coldfire.c stores the SDCLK frequency in a cookie jar
based on the JTAG ID of the ColdFire processor. We use this instead of
a compile time constant.

Once consequence of this is that the SDCLK frequency is no longer
known when init_delay() is called. We now 'hardcode' this to use
SDCLK=133 MHz and set the correct value in calibrate delay(). This is
okay because the value set in init_delay() is temporary and can be
appproximate.

Another consequence is that we can no longer cause a compile time
error if there is no SDCLK value defined. Instead, if we've
encountered an MCF54xx CPU we haven't seen before, we issue a KDEBUG
message and default SDCLK to 100 MHz.

The upside is that we don't need a separate compile target for an
M547x evaluation board running at 133 MHz.