Fix box gadget width when 3D objects are enabled
This makes window box gadgets (such as the closer) the same size as
TOS4 in all screen resolutions, which also makes the width of the
window work area in all resolutions the same as TOS4.  This is
necessary to allow 4 columns of icons in a full-screen window in
ST Low when 3D objects are enabled.