Do not validate path set by EmuCON PATH command
This makes it behave more like the Un*x PATH environment variable.
It also avoids a possible problem during startup when EmuCON sets
the path from the PATH environment variable: if the latter contained
an invalid folder specification, the whole specification was silently
ignored.  Note that this change does not affect the search for an
executable file: an invalid folder specification in the path string
is silently ignored, as before.