Fix button bug in desktop configuration dialog
When CONF_WITH_3D_OBJECTS was TRUE, the up buttons used for scrolling
through the function key list & the menu shortcut list only responded
to mouse clicks in part of their area.  The root cause of this was
the presence of an unneeded IBOX surrounding them, combined with the
expansion of objects with the FLD3D attribute.  This made ob_find()
unable to find the correct object.