Add workaround for bad sector count in bootsector
The sector count in the word at offset 0x13 in the boot sector is
used to derive the number of clusters on the disk.  For DOS-style
disks (logical sector size always 512), a value of zero indicates
that the true sector count is stored in the long at offset 0x20.
If 0x13 is zero due to a badly-formatted disk, EmuTOS will use the
(likely invalid) long value at 0x20.  In the same situation, Atari
TOS just uses the bad value and calculates a negative cluster count.
Since the error will "never" happen with hard disk partitions, we
now derive a cluster count of zero for such a diskette.