Fix ob_draw() when called for the ROOT object
This solves a regression introduced by commit 8bc21cbc. Before that commit,
when ob_draw() was called for the root of the object tree, the 'last'
variable was always set to 'NIL', regardless of the value of 'ob_next'. We
now replicate the old behavior of the get_par() function.

This fixes CodeHead Software's Icon Juggler. Thank you to Eero Tamminen for
reporting this issue.