Fix obscure filesystem bug
If multiple handles were concurrently open for writing to the same
file, clusters would be lost.  This was because each handle would
acquire free clusters independently, but the chain recorded on disk
would reflect that of the last handle closed.  This bug also exists
in Atari TOS 1-3, but not in TOS 4.