Fix bug with GDOS fonts
When GDOS is active, font_ring[2] should point to the chain of fonts
loaded by GDOS.  This was not always true, because screen() updated
it from the loaded_fonts variable in the current virtual workstation,
which was not necessarily the workstation used to load the fonts.
This error was introduced by commit 07182fe; the fix restores the
handling of font_ring[2] to the way it was before GDOS support was
removed (by commit 45c5d1c).  Thanks to Trevor B for reporting this
and providing a test program.