Rework NVDI detection during AES initialization
EmuTOS AES cannot use the vex_wheelv VDI call for mouse wheel support when
NVDI is present, since that results in an error from NVDI. The previous
NVDI detection could fail if an additional GDOS was used. The reworked
routine additionally checks for the 'NVDI' cookie.

Thank you to Philippe Noble for reporting this issue.