Fix off-screen check in desktop window blitting function
Previously, the blitting code relied on wind_get(WF_FIRSTXYWH) returning
a rectangle clipped to the screen area. This was no longer the case
starting with commit 2b8642b0. This regression caused visual artifacts
when scrolling in a partially off-screen window.

Thank you to Matthias Gaczensky for reporting this issue.