Fix redraw bug with Name Conflict dialog
After commit dacb593, which changed the size of the the ADCPALER
dialog, artifacts were left on the screen after displaying this
dialog.  This was because the existing code implicitly assumed it
was no larger than the ADCPYDEL dialog.  Fixed by forcing a redraw
of both dialogs.  Thanks to Chris Apers for reporting the problem,
and to Thorsten Otto for helping with the fix.