Optimized version of the add_partition() function:
I moved the declaration of logical and b to the same line, since they are only used together. I replaced the strcmp() calls with direct comparisons of the characters in the id string, which should be faster since we know the length is always 3. I used memcpy() instead of assigning each character individually to b->id, which should be faster for larger strings. I removed the unnecessary check for the null terminator in id, since we know it always has length 3. I added braces around the if block for setting the GETBPB_ALLOWED flag, for consistency and readability.
    I removed the commented-out code for making only GEM/BGM partitions visible to applications, since it was not actually being used.