Remove unnecessary 'repeat' for UltraSatan I/O
The call to send_command() to read the current firmware of the
UltraSatan specified 1 'repeat'.  The only purpose of repeat is to
flush the DMA FIFO for transfers of small amounts of data.  Since
the UltraSatan firmware always transfers 512 bytes for this command,
it is not necessary to flush the FIFO.