Correctly handle more than 32767 physical sectors in one Rwabs call
For file systems with differing logical and physical sector size, a single
Rwabs call can lead to more than 32767 physical sectors to be read or
written. With this commit, the buffer address is updated correctly when
such a call is therefore split up into multiple low-level IOs.

In particular, this fixes loading very large executables with at least
16 MiB combined text and data size.

Thank you to Eero Tamminen for reporting this issue and to Roger Burrows
for finding the root cause.

Co-authored-by: Roger Burrows <anodynesoftware@users.sourceforge.net>