Avoid modifying original CDB for ACSI I/O
When the CDB opcode was less than 0x1f, send_command() modified the
passed CDB by ORing in the drive code.  This did not cause problems
internally since the CDBs are built afresh each time.  However, this
will not be safe when the code is reused for SCSI driver support.