Fix bug in handling non-sector-size IDE transfers
Commit 4b44d2c changed ide_get_data()/ide_put_data() to allow IDE
data transfers to be other than a multiple of the sector size.
This did not work properly when the IDE port was addressable as a
long rather than a word.  However, this did not cause any problems
in practice since all IDE data transfers are currently multiples
of the sector size.