Fix IDE reset logic
The old code was wrong-headed: it attempted to reset each device
separately, but without selecting any specific device.  The new
code tries to follow the ATA spec more closely, by resetting once
and waiting for DRDY asserted as well as BSY de-asserted.  Thanks
to Christian Zietz for pointing out the problems in the old code.