Fix bug that caused problems in form_dial()
When the desktop exited, the pointer to the desktop resource was not
reset, even though the memory occupied by the resource was freed.
If a program allocated this memory and overwrote it, a subsequent
form_dial(FMD_FINISH) would attempt to redraw the now-overwritten
resource, which could cause a system crash or an endless loop.
The latter occurred with scummvm.
Thanks to Miro Kropacek and Eero Tamminen for reporting this.