Fix EmuCON bug when running program with -h
Under EmuCON, if you attempt to execute any program, passing it a
single '-h' argument, the command line is ignored.  This is because
'xxx -h' is converted into 'help xxx' before checking whether 'xxx'
is a built-in function, and 'help xxx' is ignored if 'xxx' is not
a built-in function.