Make Vfixmode() behave more like TOS4's version
The undocumented function Vfixmode() converts a videl mode to one
suitable for the current monitor.  If the current monitor is a VGA,
the VGA bit will be set in the converted mode.  Since the OVERSCAN
bit should be ignored if the VGA bit is set, the EmuTOS version of
Vfixmode() cleared the OVERSCAN bit.  In this situation, we now
leave the OVERSCAN bit untouched, but subsequently ignore it when
setting the mode via Vsetmode().  This is what TOS4 does.