Force alignment of start of free ST-RAM to four byte boundary
Previously, the alignment relied on the implicit settings of the toolchain
used to build EmuTOS. Now, this alignment is made explicit. For the builds
with gcc 4.6.4 binaries remain unchanged, as the alignment was already
enforced by the toolchain and the a.out format.

When built with a different toolchain, EmuTOS could previously crash on the
Falcon, where memory blocks are supposed to be aligned at four byte
boundaries since commit 79da8a5cbaff8b3e8d992c1e38169d674350f230.

Thank you to Alan Hourihane for reporting this.