Revert "Update auto-generated file for ColdFire"
This reverts commit f7f732b0eec6be5ea6cc08811ebc2277b0ae6cf3.

Remove useless change in generated file.
Key point is to commit vdi_tblit_cf.S with or after vdi_tblit.S
in order to keep the date of vdi_tblit_cf.S >= to the one of vdi_tblit.S
when someone does "git pull".
So "make" won't try to regenerate the generated file.