Fix race condition in SCC port serial transmit
The TX pending interrupt flags were not reset until after the data
was moved to the output register.  For fast data rates, another TX
interrupt could arrive before the pending interrupt was cleared,
and it would be missed, leading to a serial port hang.
Thanks to Christian Zietz for the analysis.