Fix race condition in Bconout()
This applies to all functions that implement Bconout(): bconout1(),
bconoutTT(), bconoutA(), bconoutB().  There was a small window
between checking for the register on the device being empty, and
putting the data into the iorec buffer if it wasn't.  If the device
became empty in that window, the transmit interrupt would trigger
but the interrupt wouldn't see any data in the buffer at that point,
and the data could be stuck in the iorec buffer.
Thanks to Christian Zietz for pointing this out.