Avoid some unnecessary window redraws
If a wind_set() that affects a slider is issued, the window is
redrawn, even if that call does not change the position or size of
the slider.  On the desktop, in most resolutions, this is (barely)
perceptible as a slight flicker.  However this is very noticeable
in a Falcon 16-bit display.