Fix graphical glitch in window titles
This occurs when 3D objects are available, and the name is too long
to fit in the title space.  When the title is unselected, most of the
rightmost black pixels of the closer widget are overwritten by the
white background of the window title.  This also happens in TOS 4.04.
Thanks to Thorsten Otto for reporting this problem.