EmuDesk: Fix renaming a file to replace an existing file
Previously, when dragging a file in the same folder while pressing Ctrl to
rename it, an existing file could not be specified as target. Now, this is
possible after confirming the file name. It now works the same as in Atari
TOS 2.06.

Thank you to Miro Kropáček for reporting this issue.