ELF: Mix .text and .rodata sections
This matters to ensure that all parts of the important objects
will stay at the beginning of the ROM. Otherwise, the fonts might
be put at the end.
This is particularly important for ARAnyM 1.1.0 when using a 1024k ROM
by mistake (as it only loads the first 512k). BIOS and fonts must live
in the first half of the ROM to be able to display the error message.

This change only affects ELF compilers.
It has no effect with the official m68k-atari-mint compiler using a.out.