Disable static inlines
When GCC decides to *not* inline the function, a copy of the function body
is generated in each .o file. Without LTO, such duplicated functions
can't be merged. This finally results into a bloated binary.
This actually happened 11 times in the 192k ROM.

On the other hand, actual inlining only happened a few times,
in non time-critical places.

This fix saves 52 bytes in the 192k ROM.

Note: As we don't compile with -ffreestanding, GCC can still call its
builtin functions. As it knows what the C library functions are supposed
to do, it is free to replace the calls with something better, such as
directly using the result when previsible.

This reverts commit 7156e4610089ce561e184a0f00d6dc13e7f547d8.