Rework Crazydots graphics card detection
ATI Mach32 detection is not attempted when a Crazydots graphics card is
suspected to be installed. The Crazydots always uses an ET4000 chip.

Particularly, this fixes EmuTOS on the ATW800/2 (VME variant), where EmuTOS
would suspect a Crazydots but then crash during the Mach32 detection.
Note: The following sanity check check_for_vga() correctly determines that
the ATW800/2 is not a VGA card.

Thank you to Idek for testing this fix.