Refine 'nicelines' check
Starting with commit 45c9a33 EmuTOS replaces a disabled G_STRING object
consisting only of dashes with a 'niceline' to improve the visual appearance
of menu separators.

This caused an issue when a dialog box tried to display a single dash as
disabled G_STRING object. Now, EmuTOS uses an improved heuristic for
'nicelines', only considering strings consisting only of dashes with a
minimum length of 4.

Thank you to Uwe Seimet for reporting this issue and helping with testing.