Fix SCSIDRV InquireSCSI function
According to the SCSIDRV specification, during the first call of InquireSCSI
the driver must zero the private data field, and then set the bits for the
busses it supports one after another. EmuTOS did exactly the opposite,
clearing the bits. This caused issues when multiple SCSI drivers were loaded.

This fixes #59.

Thank you to Uwe Seimet for reporting this issue.